requirements.txt
pdfplumber
pandas
openpyxl
groq
python-dotenv
tabulate